Move constants into core package
