+++ packages/crypto/Cargo.toml
@@ -12,23 +12,21 @@ license = "Apache-2.0"
 bench = false
 
 [dependencies]
-cfg-if = "1.0.0"
-digest = "0.10"
-ecdsa = "0.16.2" # Not used directly, but needed to bump transitive dependency, see: https://github.com/CosmWasm/cosmwasm/pull/1899 for details.
-ed25519-zebra = { version = "4.0.3", default-features = false }
-k256 = { version = "0.13.3", default-features = false, features = ["ecdsa"] }
-p256 = { version = "0.13.2", default-features = false, features = ["ecdsa"] }
-rand_core = "0.6"
-thiserror = "1.0.26"
-
-[target.'cfg(not(target_arch = "wasm32"))'.dependencies]
 ark-bls12-381 = "0.4.0"
 ark-ec = { version = "0.4.2", features = ["parallel"] }
 ark-ff = { version = "0.4.2", features = ["asm", "parallel"] }
 ark-serialize = "0.4.2"
+cosmwasm-core = { version = "2.0.1", path = "../core" }
+digest = "0.10"
+ecdsa = "0.16.2" # Not used directly, but needed to bump transitive dependency, see: https://github.com/CosmWasm/cosmwasm/pull/1899 for details.
+ed25519-zebra = { version = "4.0.3", default-features = false }
+k256 = { version = "0.13.3", default-features = false, features = ["ecdsa"] }
 num-traits = "0.2.18"
+p256 = { version = "0.13.2", default-features = false, features = ["ecdsa"] }
+rand_core = "0.6"
 rayon = "1.9.0"
 sha2 = "0.10"
+thiserror = "1.0.26"
 
 [dev-dependencies]
 base64 = "0.22.0"
@@ -43,6 +41,10 @@ hex = { version = "0.4", features = ["serde"] }
 hex-literal = "0.4.1"
 english-numbers = "0.3"
 glob = "0.3.1"
+
+[features]
+default = ["std"]
+std = []
 
 [[bench]]
 name = "main"
