+++ packages/crypto/benches/main.rs
@@ -15,11 +15,11 @@ use digest::Digest;
 use k256::ecdsa::SigningKey; // type alias
 use sha2::Sha256;
 
+use cosmwasm_core::{BLS12_381_G1_GENERATOR, BLS12_381_G1_POINT_LEN, BLS12_381_G2_POINT_LEN};
 use cosmwasm_crypto::{
     bls12_381_aggregate_g1, bls12_381_aggregate_g2, bls12_381_hash_to_g1, bls12_381_hash_to_g2,
     bls12_381_pairing_equality, ed25519_batch_verify, ed25519_verify, secp256k1_recover_pubkey,
     secp256k1_verify, secp256r1_recover_pubkey, secp256r1_verify, HashFunction,
-    BLS12_381_G1_GENERATOR, BLS12_381_G1_POINT_LEN, BLS12_381_G2_POINT_LEN,
 };
 use std::cmp::min;
