+++ packages/crypto/tests/bls12_381.rs
@@ -3,10 +3,10 @@ use std::{error::Error, fs};
 use ark_serialize::{CanonicalDeserialize, CanonicalSerialize};
 use base64::engine::general_purpose::STANDARD;
 use base64_serde::base64_serde_type;
+use cosmwasm_core::{BLS12_381_G1_GENERATOR, BLS12_381_G2_POINT_LEN};
 use cosmwasm_crypto::{
     bls12_381_aggregate_g1, bls12_381_aggregate_g2, bls12_381_g1_is_identity,
     bls12_381_g2_is_identity, bls12_381_hash_to_g2, bls12_381_pairing_equality, HashFunction,
-    BLS12_381_G1_GENERATOR, BLS12_381_G2_POINT_LEN,
 };
 
 const PROOF_OF_POSSESSION_DST: &[u8] = b"BLS_SIG_BLS12381G2_XMD:SHA-256_SSWU_RO_POP_";
