+++ packages/std/src/lib.rs
@@ -141,6 +141,5 @@ pub use crate::imports::{ExternalApi, ExternalQuerier, ExternalStorage};
 #[cfg(not(target_arch = "wasm32"))]
 pub mod testing;
 
-pub use cosmwasm_crypto::{BLS12_381_G1_GENERATOR, BLS12_381_G2_GENERATOR};
-
+pub use cosmwasm_core::{BLS12_381_G1_GENERATOR, BLS12_381_G2_GENERATOR};
 pub use cosmwasm_derive::entry_point;
