+++ packages/std/src/testing/mock.rs
@@ -1169,7 +1169,7 @@ mod tests {
     #[cfg(feature = "staking")]
     use crate::{Decimal, Delegation};
     use base64::{engine::general_purpose, Engine};
-    use cosmwasm_crypto::BLS12_381_G1_GENERATOR;
+    use cosmwasm_core::BLS12_381_G1_GENERATOR;
     use hex_literal::hex;
     use serde::Deserialize;
