+++ packages/vm/Cargo.toml
@@ -45,6 +45,7 @@ clru = "0.6.1"
 crc32fast = "1.3.2"
 bech32 = "0.11.0"
 # Uses the path when built locally; uses the given version from crates.io when published
+cosmwasm-core = { version = "2.0.1", path = "../core" }
 cosmwasm-std = { version = "2.0.1", path = "../std", default-features = false, features = ["std"] }
 cosmwasm-crypto = { version = "2.0.1", path = "../crypto" }
 derivative = "2"
